Impuesto, cambios carrito

+++ database/migrations/2018_02_10_214126_create_items_table.php
@@ -20,7 +20,7 @@ class CreateItemsTable extends Migration
             $table->string('producto');
             $table->longText('boletos');
             $table->integer('cantidad');
-            $table->integer('precio');
+            $table->double('precio');
             $table->dateTime('fecha');
             $table->timestamps();
